Move `AuthorizeUser` command to decidim-verifications
