+++ decidim-admin/app/commands/decidim/admin/create_managed_user.rb
@@ -45,7 +45,7 @@ module Decidim
 
       def authorized_user?
         form.authorization.user = @user
-        AuthorizeUser.call(form.authorization) do
+        Verifications::AuthorizeUser.call(form.authorization) do
           on(:ok) do
             return true
           end
