+++ decidim-verifications/spec/commmands/decidim/verifications/authorize_user_spec.rb
@@ -2,20 +2,20 @@
 
 require "spec_helper"
 
-module Decidim
+module Decidim::Verifications
   describe AuthorizeUser do
     subject { described_class.new(handler) }
 
     let(:user) { create(:user) }
     let(:document_number) { "12345678X" }
     let(:handler) do
-      DummyAuthorizationHandler.new(
+      Decidim::DummyAuthorizationHandler.new(
         document_number: document_number,
         user: user
       )
     end
 
-    let(:authorizations) { Authorizations.new(user: user, granted: true) }
+    let(:authorizations) { Decidim::Authorizations.new(user: user, granted: true) }
 
     context "when the form is not authorized" do
       before do
